+++ ReleaseNotes.md
@@ -52,6 +52,7 @@
 - cta/CTA#830 - Add script for building CTA locally
 - cta/CTA#847 - Introduced version-lock for Oracle instant client
 - cta/CTA#821 - Improved pipeline logic for faster performance and correct cancel propagation
+- cta/CTA#708 - Added a basic test for the archive metadata to the CI
 
 ### Catalogue Schema
 - cta/CTA#801 - Update CTA catalogue schema to version 15.0